Fix preloaded_vars.conf not found bug

+++ install_api.sh
@@ -296,14 +296,18 @@ setup_api() {
     if [ -d $API_PATH ]; then
         backup_api
         print ""
-        reinstall_preloaded=$(cat preloaded_vars.conf | sed -e '/^#/ d' | grep REINSTALL)
-        if [[ ! -z "$reinstall_preloaded" ]]; then
-            response=$(echo $reinstall_preloaded | cut -d'=' -f 2 | tr -d '\r')
-            case $response in
-                [nN] ) update="no";;
-                [yY] ) update="yes";;
-            esac
-        else
+        reinstall_preloaded=""
+        if [[ -f preloaded_vars.conf ]]; then
+            reinstall_preloaded=$(cat preloaded_vars.conf | sed -e '/^#/ d' | grep REINSTALL)
+            if [[ ! -z "$reinstall_preloaded" ]]; then
+                response=$(echo $reinstall_preloaded | cut -d'=' -f 2 | tr -d '\r')
+                case $response in
+                    [nN] ) update="no";;
+                    [yY] ) update="yes";;
+                esac
+            fi
+        fi
+        if [[ ! -f preloaded_vars.conf || -z "$reinstall_preloaded" ]]; then
             while true; do
                 read -p "Wazuh API is already installed (version $API_OLD_VERSION). Do you want to update it? [y/n]: " yn
                 case $yn in
@@ -313,13 +317,17 @@ setup_api() {
             done
         fi
         if [ "X${update}" == "Xno" ]; then
-            remove_preloaded=$(cat preloaded_vars.conf | sed -e '/^#/ d' | grep REMOVE)
-            if [[ ! -z "$remove_preloaded" ]]; then
-                response=$(echo $remove_preloaded | cut -d'=' -f 2 | tr -d '\r')
-                case $response in
-                    [nN] ) print "Not possible to install the API.\nExiting."; exit 1; break;;
-                esac
-            else
+            remove_preloaded=""
+            if [[ -f preloaded_vars.conf ]]; then
+                remove_preloaded=$(cat preloaded_vars.conf | sed -e '/^#/ d' | grep REMOVE)
+                if [[ ! -z "$remove_preloaded" ]]; then
+                    response=$(echo $remove_preloaded | cut -d'=' -f 2 | tr -d '\r')
+                    case $response in
+                        [nN] ) print "Not possible to install the API.\nExiting."; exit 1; break;;
+                    esac
+                fi
+            fi
+            if [[ ! -f preloaded_vars.conf || -z "$remove_preloaded" ]]; then
                 while true; do
                     print ""
                     read -p "The installation directory already exists. Should I delete it? [y/n]: " yn
